backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-filter==25.1
PyJWT==2.10.1
pytz==2025.1
sqlparse==0.5.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sendgrid==6.11.0
requests==2.32.3
